refactor admin checklist controller
